+++ packages/plugin/src/theia.d.ts
@@ -9188,6 +9188,38 @@ declare module '@theia/plugin' {
     }
 
     /**
+     * Options for starting a debug session.
+     */
+    export interface DebugSessionOptions {
+
+        /**
+         * When specified the newly created debug session is registered as a "child" session of this
+         * "parent" debug session.
+         */
+        parentSession?: DebugSession;
+
+        /**
+         * Controls whether this session should have a separate debug console or share it
+         * with the parent session. Has no effect for sessions which do not have a parent session.
+         * Defaults to Separate.
+         */
+        consoleMode?: DebugConsoleMode;
+
+        /**
+         * Controls whether this session should run without debugging, thus ignoring breakpoints.
+         * When this property is not specified, the value from the parent session (if there is one) is used.
+         */
+        noDebug?: boolean;
+
+        /**
+         * Controls if the debug session's parent session is shown in the CALL STACK view even if it has only a single child.
+         * By default, the debug session will never hide its parent.
+         * If compact is true, debug sessions with a single child are hidden in the CALL STACK view to make the tree more compact.
+         */
+        compact?: boolean;
+    }
+
+    /**
      * A debug configuration provider allows to add the initial debug configurations to a newly created launch.json
      * and to resolve a launch configuration before it is used to start a new debug session.
      * A debug configuration provider is registered via #debug.registerDebugConfigurationProvider.
@@ -9556,7 +9588,7 @@ declare module '@theia/plugin' {
          * @param nameOrConfiguration Either the name of a debug or compound configuration or a [DebugConfiguration](#DebugConfiguration) object.
          * @return A thenable that resolves when debugging could be successfully started.
          */
-        export function startDebugging(folder: WorkspaceFolder | undefined, nameOrConfiguration: string | DebugConfiguration): PromiseLike<boolean>;
+        export function startDebugging(folder: WorkspaceFolder | undefined, nameOrConfiguration: string | DebugConfiguration, options: DebugSessionOptions): PromiseLike<boolean>;
 
         /**
          * Add breakpoints.